# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/health": {
            "get": {
                "description": "get application health status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Status"
                ],
                "summary": "Application health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ApplicationStatus"
                        }
                    }
                }
            }
        },
        "/jobs": {
            "get": {
                "description": "Query job status with identity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Query Job Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "job service type",
                        "name": "service",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job task type",
                        "name": "task",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job domain type",
                        "name": "domain",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Job"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a job with specified SPEC",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Create Job",
                "parameters": [
                    {
                        "description": "body for create a job",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.CreateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/common.Job"
                        }
                    }
                }
            }
        },
        "/jobs/logs": {
            "get": {
                "description": "Query job status with identity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "Query Job Logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "job service type",
                        "name": "service",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job task type",
                        "name": "task",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job domain type",
                        "name": "domain",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job StepID",
                        "name": "stepID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Step log start time",
                        "name": "startTimeUUID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "max record to retrieve",
                        "name": "maxRecord",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "app.ApplicationInfo": {
            "type": "object",
            "properties": {
                "commitID": {
                    "type": "string"
                },
                "releaseAt": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
        "application.CreateJobRequest": {
            "type": "object",
            "required": [
                "domain",
                "engine",
                "service",
                "spec",
                "task",
                "userID"
            ],
            "properties": {
                "domain": {
                    "type": "string"
                },
                "engine": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                },
                "spec": {
                    "type": "object",
                    "additionalProperties": true
                },
                "task": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "common.Job": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "endTime": {
                    "type": "string"
                },
                "engine": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "service": {
                    "type": "string"
                },
                "spec": {
                    "type": "object",
                    "additionalProperties": true
                },
                "startTime": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.Step"
                    }
                },
                "task": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "common.Step": {
            "type": "object",
            "properties": {
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "controller.ApplicationStatus": {
            "type": "object",
            "properties": {
                "info": {
                    "$ref": "#/definitions/app.ApplicationInfo"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}